score 부분 prresult에 함수 추가
