+++ apps/preresult/views.py
@@ -2,7 +2,8 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.utils import timezone
 from django.urls import reverse
-from apps.group.models import Group, MemberState, Idea
+from django.db.models import Q 
+from apps.group.models import Group, MemberState, Idea, Vote 
 from apps.group.views import State, TeamNumber, redirect_by_auth 
 
 # Create your views here.
@@ -110,3 +111,24 @@ def preresult_modify(request, group_id):
     else:
         redirect_url = reverse("group:group_detail", kwargs={"group_id": group_id})
         return redirect(redirect_url)
+
+
+def calculate_idea_scores(group_id):
+    ideas = Idea.objects.filter(group_id=group_id)
+
+    for idea in ideas:
+        votes_count = Vote.objects.filter(
+            group_id=group_id
+        ).filter(
+            Q(idea_vote1=idea) | 
+            Q(idea_vote2=idea) | 
+            Q(idea_vote3=idea)
+        ).count()
+        
+        idea.votes = votes_count
+        idea.save()
+
+# preresult부분에서 calculate_idea_scores 함수를 호출하여 
+# 특정 그룹의 모든 아이디어에 대한 점수를 계산하는 방식으로 생각했습니다.
+# 아래는 예시
+calculate_idea_scores(1) #group_id가 1인 그룹의 아이디어 점수를 계산하는 경우
